@@ -1,7 +1,6 @@
-module.exports = {
+export default {
   plugins: {
     'postcss-import': {},
-    'tailwindcss/nesting': {},
     tailwindcss: {},
     autoprefixer: {},
   },
