@@ -4,7 +4,7 @@ import {
   faPlaneDeparture,
 } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import React, { useState } from 'react'
+import { useState } from 'react'
 import Button from './Button'
 
 const ExampleIcon = (
