@@ -1,6 +1,6 @@
 import { faX } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import React, { useState } from 'react'
+import { useState } from 'react'
 
 import Button from './Button'
 import Checkbox from './Checkbox'
