@@ -1,5 +1,4 @@
 import { Form, Formik } from 'formik'
-import React from 'react'
 import * as yup from 'yup'
 import Button from '../Button'
 import FormikSwitchField from './FormikSwitchField'
@@ -41,7 +40,7 @@ export const Disabled = () => {
           switch: false,
         }}
         isInitialValid={false}
-        onSubmit={(values) => {
+        onSubmit={() => {
           alert('Switch is disabled')
         }}
       >
