@@ -1,6 +1,6 @@
 import { faEye, faEyeSlash } from '@fortawesome/free-solid-svg-icons'
 import { Form, Formik } from 'formik'
-import React, { useState } from 'react'
+import { useState } from 'react'
 import * as yup from 'yup'
 import Button from '../Button'
 import FormikTextField from './NewFormikTextField'
