@@ -1,42 +1,41 @@
 import './tailwind.css'
 
-export * from './Button'
-export * from './Checkbox'
-export * from './Collapsible'
-export * from './ColorPicker'
-export * from './Countdown'
-export * from './CycleCountdown'
-export * from './CycleProgress'
-export * from './DateChanger'
-export * from './Dropdown'
-export * from './Header'
-export * from './Modal'
-export * from './Navigation'
-export * from './NotificationBadgeWrapper'
-export * from './Progress'
-export * from './Prose'
-export * from './Select'
-export * from './Slider'
-export * from './StepProgress'
-export * from './Switch'
-export * from './Table'
-export * from './Tabs'
+export { default as Button } from './Button'
+export { default as Checkbox } from './Checkbox'
+export { default as Collapsible } from './Collapsible'
+export { default as ColorPicker } from './ColorPicker'
+export { default as Countdown } from './Countdown'
+export { default as CycleCountdown } from './CycleCountdown'
+export { default as CycleProgress } from './CycleProgress'
+export { default as DateChanger } from './DateChanger'
+export { default as Dropdown } from './Dropdown'
+export { default as Header } from './Header'
+export { default as Modal } from './Modal'
+export { default as Navigation } from './Navigation'
+export { default as NotificationBadgeWrapper } from './NotificationBadgeWrapper'
+export { default as Progress } from './Progress'
+export { default as Prose } from './Prose'
+export { default as Select } from './Select'
+export { default as Slider } from './Slider'
+export { default as StepProgress } from './StepProgress'
+export { default as Switch } from './Switch'
+export { default as Table } from './Table'
 export { default as Tabs } from './Tabs'
-export * from './Tag'
-export * from './Toast'
-export * from './Tooltip'
-export * from './UserNotification'
-export * from './Workflow'
-export * from './forms/FormikColorPicker'
-export * from './forms/FormikDateChanger'
-export * from './forms/FormikDateField'
-export * from './forms/FormikNumberField'
-export * from './forms/FormikPinField'
-export * from './forms/FormikSelectField'
-export * from './forms/FormikSwitchField'
-export * from './forms/FormikTextField'
-export * from './forms/FormikTextareaField'
-export * from './forms/Label'
+export { default as Tag } from './Tag'
+export { default as Toast } from './Toast'
+export { default as Tooltip } from './Tooltip'
+export { default as UserNotification } from './UserNotification'
+export { default as Workflow } from './Workflow'
+export { default as FormikColorPicker } from './forms/FormikColorPicker'
+export { default as FormikDateChanger } from './forms/FormikDateChanger'
+export { default as FormikDateField } from './forms/FormikDateField'
+export { default as FormikNumberField } from './forms/FormikNumberField'
+export { default as FormikPinField } from './forms/FormikPinField'
+export { default as FormikSelectField } from './forms/FormikSelectField'
+export { default as FormikSwitchField } from './forms/FormikSwitchField'
+export { default as FormikTextField } from './forms/FormikTextField'
+export { default as FormikTextareaField } from './forms/FormikTextareaField'
+export { default as Label } from './forms/Label'
 export { default as NewFormikDateChanger } from './forms/NewFormikDateChanger'
 export { default as NewFormikDateField } from './forms/NewFormikDateField'
 export { default as NewFromikNumberField } from './forms/NewFormikNumberField'
@@ -45,8 +44,8 @@ export { default as NewFormikSelectField } from './forms/NewFormikSelectField'
 export { default as NewFormikSwitchField } from './forms/NewFormikSwitchField'
 export { default as NewFormikTextField } from './forms/NewFormikTextField'
 export { default as NewFormikTextareaField } from './forms/NewFormikTextareaField'
-export * from './forms/NumberField'
-export * from './forms/SelectField'
-export * from './forms/TextField'
-export * from './forms/TextareaField'
-export * from './hooks/useArrowNavigation'
+export { default as NumberField } from './forms/NumberField'
+export { default as SelectField } from './forms/SelectField'
+export { default as TextField } from './forms/TextField'
+export { default as TextareaField } from './forms/TextareaField'
+export { useArrowNavigation } from './hooks/useArrowNavigation'
